requirements.txt
Django==1.10.1
django-jinja==2.2.1
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.2
python-dateutil==2.5.3
six==1.10.0
